+++ executor/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "executor"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+plonky2 = { git = "https://github.com/zkMIPS/plonky2.git", branch = "zkm_dev" }
+plonky2_util = { git = "https://github.com/zkMIPS/plonky2.git", branch = "zkm_dev" }
+plonky2_maybe_rayon = { git = "https://github.com/zkMIPS/plonky2.git", branch = "zkm_dev" }
+zkm = { git = "https://github.com/zkMIPS/zkm.git", branch = "main" }
+
+itertools = "0.11.0"
+log = { version = "0.4.14", default-features = false }
+anyhow = "1.0.75"
+num = "0.4.0"
+num-bigint = "0.4.3"
+serde = { version = "1.0.144", features = ["derive"] }
+serde_json = "1.0"
+tiny-keccak = "2.0.2"
+rand = "0.8.5"
+rand_chacha = "0.3.1"
+once_cell = "1.13.0"
+static_assertions = "1.1.0"
+keccak-hash = "0.10.0"
+byteorder = "1.5.0"
+hex = "0.4"
+hashbrown = { version = "0.14.0", default-features = false, features = ["ahash", "serde"] } # NOTE: When upgrading, see `ahash` dependency.
+lazy_static = "1.4.0"
+elf = { version = "0.7", default-features = false }
