+++ service/Cargo.toml
@@ -8,6 +8,7 @@ edition = "2021"
 [dependencies]
 prover = { path = "../prover" }
 stage = {path = "../stage"}
+executor = {path = "../executor"}
 tonic = "0.8.1"
 prost = "0.11.0"
 tokio = { version = "1.21.0", features = ["macros", "rt-multi-thread", "signal"] }
